refactor(parsers): split skill audit behavior helpers
